+++ src/engine/world_builder/overlaymap.h
@@ -53,7 +53,7 @@ overlaymap_set_tile_texture(
     int x,
     int z,
     int level,
-    uint8_t texture_id,
+    int16_t texture_id,
     uint16_t texture_avg_hsl16);
 
 struct OverlaymapTile*
